daqdev/NSCLDAQ#1028 - MAKE MPI COMPILATIONS TRANSPARENT
This commit addresses the need to provide CXX=(mpi compiler) for
openMPI at least.  That compiler has a -showme switch which
provides the g++ command it fires off.  We cut g++ from that
and provide the MPI_FLAGS variable which is then added to the
compilation and the link flags.

This change allows us to use the standard compilers and not have
to provide a CXX override on configure.