FIX ERRORS IN IdleTimer
The flush should be unconditional rather thatn depend on XON
somewhere.  That will, if there are empty queues, result in a
timed flush eventually if the XOFF is set and, if XONed, new
data may eventually come in.

The key point is that only this flush of queues can eventually
cause XON's to happen.  Every case I can see where XON is used
to considere whether or not to flush results in hangs that are
deadlocked on XOFF.