daqdev/NSCLDAQ#898 - MCFD16 GIVES INVALID MODE IF BLW0 IS SELECTED IN FP
Given that we're pushing people over into the mesytec control
program and the RC bus pod; a kludgy respones is probabl fine.
Specificaly we only look at the bottom bit of the mode register
regardless of the front panel.  The assumption is that any additional
bits are garbage and should be ignored.